em: add external editor
add an option for an external editor. should work on Wayland and buggy
X11s. can be enabled by setting: editor.external_editor : true, and
specifying an editor cmd which only takes one argument (%1) for the
file, _and_ runs in the _forground_ (like before) so that we can monitor
the process.